pyproject.toml
[project]
name = "moonshine-flow"
dynamic = ["version"]
description = "Push-to-talk transcription daemon for macOS using Moonshine"
readme = "README.md"
requires-python = ">=3.11,<3.12"
authors = [
  { name = "Moonshine Flow Contributors" }
]
dependencies = [
  "numpy>=2.0.0",
  "sounddevice>=0.5.1",
  "pynput>=1.7.7",
  "pyperclip>=1.9.0",
  "websockets>=12.0",
  "mlx-whisper>=0.4.2; platform_system == 'Darwin' and platform_machine == 'arm64'",
  "voxmlx>=0.0.2; platform_system == 'Darwin' and platform_machine == 'arm64'",
  "transformers>=4.48.0",
  "mistral-common[audio]>=1.5.0",
  "pydantic>=2.9.0",
  "tomli-w>=1.1.0",
  "moonshine-voice>=0.0.49",
  "pyobjc-core>=11.0; platform_system == 'Darwin'",
  "pyobjc-framework-Quartz>=11.0; platform_system == 'Darwin'",
  "pyobjc-framework-AVFoundation>=11.0; platform_system == 'Darwin'",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-mock>=3.14.0",
  "ruff>=0.8.0",
]

[project.scripts]
moonshine-flow = "moonshine_flow.cli:main"
mflow = "moonshine_flow.cli:main"

[build-system]
requires = ["hatchling>=1.27.0", "hatch-vcs>=0.4.0"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "vcs"

[tool.hatch.version.raw-options]
tag_regex = "^v(?P<version>.+)$"
version_scheme = "no-guess-dev"
local_scheme = "no-local-version"
fallback_version = "0.0.0.dev0"

[tool.uv]
default-groups = []

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["B008"]
